+++ victoria.core/src/core/object/command_queue.cpp
@@ -7,31 +7,35 @@
 
 Error CommandQueue::push_commandp(const CallableMethod &p_method, const Variant **p_args, int p_argcount) {
 	// Allocate a new page if needed
-	if (bytes_per_page[allocated_memory.page_count - 1] < sizeof(Message)) {
-		uint8_t *new_memory = (uint8_t *)Memory::vreallocate(allocated_memory.mem,
-															 ++allocated_memory.page_count * PageAllocator::PAGE_SIZE);
+	if (PAGE_SIZE - bytes_per_page[allocated_memory.page_count - 1] < sizeof(Message)) {
+		uint8_t *new_memory =
+			(uint8_t *)Memory::vreallocate(allocated_memory.mem, ++allocated_memory.page_count * PAGE_SIZE);
 		ERR_COND_NULL_R(new_memory, ERR_OUT_OF_MEMORY);
 		allocated_memory.mem = new_memory;
 		allocated_memory.page_offset = 0;
-		bytes_per_page.push_back(0);
+		if (bytes_per_page.size() < allocated_memory.page_count) {
+			bytes_per_page.push_back(0);
+		}
 	}
 	// Get next possible location on the heap
 	uint32_t &offset = bytes_per_page[allocated_memory.page_count - 1];
-	uint8_t *mem = allocated_memory.mem + (allocated_memory.page_count * PageAllocator::PAGE_SIZE);
+	uint8_t *mem = allocated_memory.mem + (allocated_memory.page_count - 1) * PAGE_SIZE;
 
 	Message *mptr = vnew_placement(mem + offset, Message);
 	offset += sizeof(Message);
 	mptr->callable = p_method;
 	mptr->argcount = p_argcount;
 
-	if (PageAllocator::PAGE_SIZE - offset < sizeof(Variant) * p_argcount) {
-		uint8_t *new_memory = (uint8_t *)Memory::vreallocate(allocated_memory.mem,
-															 ++allocated_memory.page_count * PageAllocator::PAGE_SIZE);
+	if (PAGE_SIZE - offset < sizeof(Variant) * p_argcount) {
+		uint8_t *new_memory =
+			(uint8_t *)Memory::vreallocate(allocated_memory.mem, ++allocated_memory.page_count * PAGE_SIZE);
 		ERR_COND_NULL_R(new_memory, ERR_OUT_OF_MEMORY);
 		allocated_memory.mem = new_memory;
 		allocated_memory.page_offset = 0;
-		bytes_per_page.push_back(0);
-		mem = allocated_memory.mem + PageAllocator::PAGE_SIZE * (allocated_memory.page_count - 1);
+		if (bytes_per_page.size() < allocated_memory.page_count) {
+			bytes_per_page.push_back(0);
+		}
+		mem = allocated_memory.mem + PAGE_SIZE * (allocated_memory.page_count - 1);
 		offset = bytes_per_page[allocated_memory.page_count - 1];
 	}
 
@@ -47,24 +51,27 @@ Error CommandQueue::push_commandp(const CallableMethod &p_method, const Variant 
 }
 
 void CommandQueue::flush() {
-	int page = 0;
-	int offset = 0;
+	uint32_t page = 0;
+	uint32_t offset = 0;
 	while (page < allocated_memory.page_count && offset < bytes_per_page[page]) {
-		Message *m = (Message *)allocated_memory.mem + (page * PageAllocator::PAGE_SIZE) + offset;
+		Message *m = (Message *)(allocated_memory.mem + (page * PAGE_SIZE) + offset);
 		offset += sizeof(Message);
 		// Find arguments
 		uint32_t varmem = m->argcount * sizeof(Variant);
-		if (PageAllocator::PAGE_SIZE - offset < varmem) {
+		if (PAGE_SIZE - offset < varmem) {
 			page++;
 			offset = 0;
 		}
 
-		Variant *args = (Variant *)(allocated_memory.mem + (PageAllocator::PAGE_SIZE * page) + offset);
+		Variant *args = (Variant *)(allocated_memory.mem + (PAGE_SIZE * page) + offset);
 
 		// Construct argptrs
-		Variant **argptrs = (Variant **)alloca(sizeof(Variant *) * m->argcount);
-		for (int i = 0; i < m->argcount; i++) {
-			argptrs[i] = &args[i];
+		Variant **argptrs = nullptr;
+		if (m->argcount > 0) {
+			argptrs = (Variant **)alloca(sizeof(Variant *) * m->argcount);
+			for (int i = 0; i < m->argcount; i++) {
+				argptrs[i] = &args[i];
+			}
 		}
 
 		Error err;
@@ -90,11 +97,58 @@ void CommandQueue::flush() {
 	allocated_memory.page_count = 1;
 }
 
-void CommandQueue::clear() {}
+void CommandQueue::clear() {
+	uint32_t page = 0;
+	uint32_t offset = 0;
+	while (page < allocated_memory.page_count && offset < bytes_per_page[page]) {
+		Message *m = (Message *)(allocated_memory.mem + (page * PAGE_SIZE) + offset);
+		offset += sizeof(Message);
+		if (offset < sizeof(Variant) * m->argcount) {
+			page++;
+			offset = 0;
+		}
+		Variant *args = (Variant *)(allocated_memory.mem + (page * PAGE_SIZE) + offset);
+		for (int i = 0; i < m->argcount; i++) {
+			args[i].~Variant();
+			offset += sizeof(Variant);
+		}
 
-CommandQueue::CommandQueue() {
-	allocated_memory.mem = (uint8_t *)Memory::vallocate_zeroed(PageAllocator::PAGE_SIZE);
+		m->~Message();
+
+		if (offset == bytes_per_page[page]) {
+			page++;
+			offset = 0;
+		}
+	}
+
+	bytes_per_page[0] = 0;
 	allocated_memory.page_count = 1;
 }
 
-CommandQueue::~CommandQueue() {}
+CommandQueue::CommandQueue() {
+	allocated_memory.mem = (uint8_t *)Memory::vallocate_zeroed(PAGE_SIZE);
+	allocated_memory.page_count = 1;
+	bytes_per_page.push_back(0);
+}
+
+CommandQueue::~CommandQueue() {
+	clear();
+
+	if (allocated_memory.mem) {
+		Memory::vfree(allocated_memory.mem);
+	}
+}
+
+GlobalCommandQueue *GlobalCommandQueue::singleton = nullptr;
+
+GlobalCommandQueue *GlobalCommandQueue::get_singleton() {
+	return singleton;
+}
+
+GlobalCommandQueue::GlobalCommandQueue() {
+	singleton = this;
+}
+
+GlobalCommandQueue::~GlobalCommandQueue() {
+	singleton = nullptr;
+}