+++ victoria.runtime/src/scene/gui/colour_rect.cpp
@@ -16,5 +16,5 @@ Vector4 ColourRect::get_colour() const {
 void ColourRect::set_colour(const Vector4 &p_colour) {
 	colour = p_colour;
 
-	force_redraw();
+	queue_redraw();
 }